requirements.txt
Flask
Flask-Login
flask-session
psycopg2-binary
python-dotenv
requests
werkzeug